requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
newspaper3k==0.2.8
transformers==4.45.2
torch>=2.3.0,<3.0.0
streamlit==1.38.0
